adding ressources + units tests

+++ LittleConqueror.AppService/Domain/DrivingModels/Commands/CreateRegistrationLinkCommand.cs
@@ -3,6 +3,6 @@ namespace LittleConqueror.AppService.Domain.DrivingModels.Commands;
 public class CreateRegistrationLinkCommand
 {
     public string Role { get; set; }
-    public int FirstOsmId { get; set; }
+    public long FirstOsmId { get; set; }
     public char FirstOsmType { get; set; }
 }
