+++ LittleConqueror.AppService/Domain/Models/Entities/Base/Entity.cs
@@ -1,3 +1,3 @@
 namespace LittleConqueror.AppService.Domain.Models.Entities.Base;
 
-public abstract class Entity : EntityBase<int>;
+public abstract class Entity : EntityBase<long>;
