+++ LittleConqueror.AppService/Domain/Models/RegistrationLinkData.cs
@@ -4,6 +4,6 @@ public class RegistrationLinkData
 {
     public bool Valid { get; set; }
     public string Role { get; set; }
-    public int FirstOsmId { get; set; }
+    public long FirstOsmId { get; set; }
     public char FirstOsmType { get; set; }
 }
