+++ LittleConqueror.AppService/DrivenPorts/Specifications/TerritoryFromUserIdWithCitiesSpec.cs
@@ -5,7 +5,7 @@ namespace LittleConqueror.AppService.DrivenPorts.Specifications;
 public sealed class TerritoryFromUserIdWithCitiesSpec 
     : BaseSpecification<Territory>
 {
-    public TerritoryFromUserIdWithCitiesSpec(int userId) 
+    public TerritoryFromUserIdWithCitiesSpec(long userId) 
         : base(t => t.OwnerId == userId)
     {
         AddInclude(t => t.Cities);
